BACKEND/requirements.txt
fastapi
uvicorn
pydantic
langchain-google-genai
python-dotenv
pymongo
dnspython